fix: [AICPS-55-dev] Refactor code
